\name{survey.skip-package}
\alias{survey.skip-package}
\alias{survey.skip}
\docType{package}
\title{
\packageTitle{survey.skip}
}
\description{
\packageDescription{survey.skip}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{survey.skip}
\packageIndices{survey.skip}
~~ An overview of how to use the package, ~~
~~ including the most important functions ~~
}
\author{
\packageAuthor{survey.skip}

Maintainer: \packageMaintainer{survey.skip}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per ~~
~~ line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
